add non-operational periods for ports
