+++ wombat/core/service_equipment.py
@@ -6,7 +6,6 @@ servicing equipment.
 # TODO: NEED A SPECIFIC STARTUP METHOD
 from __future__ import annotations
 
-import datetime
 from copy import deepcopy
 from math import ceil
 from typing import TYPE_CHECKING, Any, Optional, Generator  # type: ignore
@@ -314,58 +313,6 @@ class ServiceEquipment(RepairsMixin):
             self.onsite = True
         self.at_system = True if set_current is not None else False
         self.current_system = set_current
-
-    def hours_to_next_operational_date(
-        self,
-        start_search_date: datetime.datetime | datetime.date | pd.Timestamp,
-        exclusion_days: int = 0,
-    ) -> float:
-        """Calculates the number of hours until the first date that is not a part of
-        the ``non_operational_dates`` given a starting datetime and for search criteria.
-        Optionally, ``exclusion_days`` can be used to account for a mobilization period
-        so that mobilization can occur during the non operational period.
-
-        Parameters
-        ----------
-        start_search_date : datetime.datetime | datetime.date | pd.Timestamp
-            The first date to be considered in the search.
-        exclusion_days : int, optional
-            The number of days to subtract from the next available datetime that
-            represents a non operational action that can occur during the non
-            operational period, such as mobilization, by default 0.
-
-        Returns
-        -------
-        float
-            The total number of hours until the next operational date.
-        """
-        current = self.env.simulation_time
-
-        if isinstance(start_search_date, datetime.date):
-            start_search_date = datetime.datetime.combine(
-                start_search_date, datetime.datetime.max.time()
-            )
-
-        # Get the forecast and filter out dates prior to the start of the search
-        dates, *_ = self.env.weather_forecast(8760)  # get the dates for the next year
-        dates = dates[np.where(dates > start_search_date)[0]]
-
-        # Find the next operational date, and if no available dates, return the time
-        # until the end of the forecast period
-        assert isinstance(self.settings, UnscheduledServiceEquipmentData)  # mypy helper
-        date_diff = set(dates.date).difference(self.settings.non_operational_dates_set)
-        if not date_diff:
-            diff = ((max(dates) - current).days + 1) * HOURS_IN_DAY
-            return diff
-        next_available = min(date_diff)
-        next_available = dates[np.where(dates.date == next_available)[0][0]]
-
-        # Compute the difference between the current time and the future date
-        diff = next_available - current
-        hours_to_next = diff.days * HOURS_IN_DAY + diff.seconds / 60 / 60
-        hours_to_next -= exclusion_days * HOURS_IN_DAY
-        hours_to_next = max(0, hours_to_next)
-        return hours_to_next
 
     def _weather_forecast(
         self, hours: int | float, which: str
@@ -1000,7 +947,7 @@ class ServiceEquipment(RepairsMixin):
         if delay == -1:
             if start == "site":
                 kw = deepcopy(kwargs)
-                kw.update({"reason": "Insufficient weather window, returne to port"})
+                kw.update({"reason": "Insufficient weather window, return to port"})
                 yield self.env.process(self.travel("site", "port", **kw))
                 yield self.env.timeout(HOURS_IN_DAY * 4)
                 yield self.env.process(
